Modify default pilatus num_images pv for fly scanning (#67)
While other detectors can have num_images set to 0, and understand this to mean they should take images indefinitely, this is not the case for the pilatus. Previously it was assumed the maximal number for a pilatus hdf writer could be set based on the limits of a int32, however it's appeared this is not the case.

With a value of 999,999 it seems to work.