Modify error handling for timed out signals when connecting.
This commit puts responsibility on anything with a .connect method
to raise its own ConnectionTimeoutError, to be then picked up by
the DeviceCollector, or not. This simplifies the complex error
handling requiring timeouts to be handled at the asyncio level and
passed up the called via (now redefined) NotConnected exceptions.