Move methods from flyer to standard detector to support multiple detector collection (#116)
This PR removes the responsibility of flyer to manage all of the detectors and instead delegates the responsibility of each detector back to themselves. This means standard detectors are now flyable- they have a kick off and complete, and can be collected. Detectors can now be collected as a group, so collect can be called on a grouping of them provided they all collect under one stream. The flyer now is only responsible for Trigger logic.

There is currently a defined order of calling prepare, flyer first then detectors, as the detectors need the trigger information provided by the flyer. However, we are looking as possibly moving this logic also up to the plan level in future changes.