Reduce boilerplate in StandardReadable
This reduces the amount of duplication and repetition when adding
signals to a StandardReadable.

As part of this, classes defining the types of signal have been created,
which control the behaviour of the Signal being registered

Signals must be registered either using the "add_children_as_readables"
contextmanager, or the "add_readables" function.