Dtype str in desc (#438)
As required by `event-model`, the `dtype` in `describe()` is a JSONschema data type, such as `"number"` or `"array"`. For Tiled (and perhaps for other consumers) we need to know that actual machine data type, such as `"<f8"` or `"<i4"`. The event-model does not currently specify a place to put this, but databroker+tiled have been using a new key `dtype_str` to place this (optionally) alongside the JSONschema `dtype`.

---------

Co-authored-by: Max Rakitin <mrakitin@bnl.gov>
Co-authored-by: Max Rakitin <mrakitin@users.noreply.github.com>
Co-authored-by: Tom C (DLS) <101418278+coretl@users.noreply.github.com>