Refactor StandardDetector kickoff method to track the number of completable frames
This commit refactors the kickoff method in the StandardDetector class to track the number of frames that can be completed after kickoff. The frames_completed attribute has been replaced with the completable_frames attribute, which represents the number of frames that can be completed. Additionally, the frames_completed attribute has been removed from the complete method and replaced with the completable_frames attribute. This change ensures that the completable_frames attribute is reset to 0 once the target number of iterations is reached and the detector is disarmed.